# Drill 6735: Implement Semi-Join for the Hash-Join operator