--- a/exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/NestedLoopJoinPOP.java
+++ b/exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/NestedLoopJoinPOP.java
@@ -40,7 +40,7 @@ public class NestedLoopJoinPOP extends AbstractJoinPop {
       @JsonProperty("joinType") JoinRelType joinType,
       @JsonProperty("condition") LogicalExpression condition
   ) {
-    super(left, right, joinType, condition, null);
+    super(left, right, joinType, false, condition, null);
   }
 
   @Override