… exec/java-exec/src/main/java/org/apache/drill/exec/planner/RuleInstance.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/HashJoinPrel.java
… exec/java-exec/src/main/resources/drill-module.conf
… exec/java-exec/src/test/java/org/apache/drill/exec/physical/impl/common/HashPartitionTest.java
… exec/java-exec/src/test/java/org/apache/drill/exec/physical/impl/join/TestBuildSidePartitioningImpl.java
… exec/java-exec/src/test/java/org/apache/drill/exec/physical/impl/join/TestHashJoinHelperSizeCalculatorImpl.java
… exec/java-exec/src/test/java/org/apache/drill/exec/physical/impl/join/TestPostBuildCalculationsImpl.java