@Ben-Zvi 2018-11-13
Good suggestions. The PR was changed - removed the 2nd and 3rd commits, and added a new commit with the Unused Hash Join Helper returning size zero.
Also created DRILL-6845 to handle the duplicates (i.e., eventually the 2nd + 3rd commits would go there).